+++ R/plot-prism.R
@@ -257,8 +257,6 @@ plot_scatter_with_corr_panel <- function(dt_response,
     Y_dt <- dt_response[, c(cellline_name, selected_metric), with = FALSE]
     tab_plot <- Y_dt[X_dt, on = .(CellLineName = CCLEName), nomatch = NULL]
 
-    tab_plot_all <- data.table::data.table()
-
     if (sum(is.na(selected_feats)) > 1) {
       tmp <- data.table::data.table(selected_feats)
       tmp[, N := seq_len(.N), by = selected_feats]
@@ -267,8 +265,10 @@ plot_scatter_with_corr_panel <- function(dt_response,
     }
 
     feat_lbl_levels <- selected_feats
+    plot_parts <- vector("list", length(selected_feats))
 
-    for (selected_feat in selected_feats) {
+    for (fi in seq_along(selected_feats)) {
+      selected_feat <- selected_feats[[fi]]
 
       if (selected_feat %chin% available_feats) {
         # remove NA
@@ -285,17 +285,18 @@ plot_scatter_with_corr_panel <- function(dt_response,
           correlation <- sqrt(r_squared)
           # add label for points driving the correlation
           dist_cooks <- sort(stats::cooks.distance(fit), decreasing = TRUE)
-          top_driving_corr <- as.numeric(names(dist_cooks)[seq_len(5)])
-          tab_plot_sel$label <- ""
-          tab_plot_sel[top_driving_corr, ]$label <- tab_plot_sel[top_driving_corr, ][[cellline_name]]
-          tab_plot_sel$col <- "no"
-          tab_plot_sel[top_driving_corr, ]$col <- "yes"
+          top_driving_corr <- as.integer(names(dist_cooks)[seq_len(5)])
+          data.table::set(tab_plot_sel, j = "label", value = "")
+          data.table::set(tab_plot_sel, i = top_driving_corr, j = "label",
+                          value = tab_plot_sel[[cellline_name]][top_driving_corr])
+          data.table::set(tab_plot_sel, j = "col", value = "no")
+          data.table::set(tab_plot_sel, i = top_driving_corr, j = "col", value = "yes")
           tab_plot_sel$feat_lbl <- selected_feat
           data.table::setnames(tab_plot_sel, selected_feat, "feat_val")
         } else {
           # dummy data when all data is NA
           feat_lbl <- paste0(selected_feat, ": all NAs")
-          feat_lbl_levels[which(selected_feats == selected_feat)] <- feat_lbl
+          feat_lbl_levels[fi] <- feat_lbl
 
           tab_plot_sel <- data.table::data.table(
             cellline_name = "",
@@ -312,7 +313,7 @@ plot_scatter_with_corr_panel <- function(dt_response,
       } else {
         # dummy data required for faceting
         feat_lbl <- paste0(selected_feat, ": all NAs")
-        feat_lbl_levels[which(selected_feats == selected_feat)] <- feat_lbl
+        feat_lbl_levels[fi] <- feat_lbl
 
         tab_plot_sel <- data.table::data.table(
           cellline_name = "",
@@ -326,8 +327,9 @@ plot_scatter_with_corr_panel <- function(dt_response,
                              old = c("cellline_name", "selected_metric"),
                              new = c(cellline_name, selected_metric))
       }
-      tab_plot_all <- rbind(tab_plot_all, tab_plot_sel)
+      plot_parts[[fi]] <- tab_plot_sel
     }
+    tab_plot_all <- data.table::rbindlist(plot_parts, fill = TRUE)
     # order vis as in selected_feats
     tab_plot_all$feat_lbl <- factor(tab_plot_all$feat_lbl, levels = feat_lbl_levels)
 
@@ -400,20 +402,15 @@ plot_boxplot_num <- function(dt_response,
   checkmate::assert_names(names(dt_depmap), must.include = c("CCLEName", selected_feat))
   checkmate::assert_string(selected_feat_meta_col, null.ok = TRUE)
   boxplot_fill <-
-    gDRutils::get_settings_from_json("BOXPLOT_FILL",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("BOXPLOT_FILL")
   hline_color <-
-    gDRutils::get_settings_from_json("HLINE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("HLINE_COLOR")
   hline_color <-
-    gDRutils::get_settings_from_json("HLINE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("HLINE_COLOR")
   jitter_poinst_color <-
-    gDRutils::get_settings_from_json("JITTER_POINST_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("JITTER_POINST_COLOR")
   edge_color <-
-    gDRutils::get_settings_from_json("EDGE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("EDGE_COLOR")
 
   selected_metric <- setdiff(names(dt_response),
                              c(cellline_name, "rId", "cId"))
@@ -497,17 +494,13 @@ plot_boxplot_num_panel <- function(dt_response,
   checkmate::assert_string(selected_feat_meta_col, null.ok = TRUE)
   checkmate::assert_int(ncol, lower = 1, null.ok = TRUE) # nolint undesirable_function_linter.
   boxplot_fill <-
-    gDRutils::get_settings_from_json("BOXPLOT_FILL",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("BOXPLOT_FILL")
   hline_color <-
-    gDRutils::get_settings_from_json("HLINE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("HLINE_COLOR")
   jitter_poinst_color <-
-    gDRutils::get_settings_from_json("JITTER_POINST_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("JITTER_POINST_COLOR")
   edge_color <-
-    gDRutils::get_settings_from_json("EDGE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("EDGE_COLOR")
 
   selected_metric <- setdiff(names(dt_response),
                              c(cellline_name, "rId", "cId"))
@@ -539,8 +532,6 @@ plot_boxplot_num_panel <- function(dt_response,
     ls_lbl <- unique(as.vector(as.matrix(tab_plot[, -c(cellline_name, selected_metric), with = FALSE])))
     dummy_feat_val <- ls_lbl[!is.na(ls_lbl)]
 
-    tab_plot_all <- data.table::data.table()
-
     if (sum(is.na(selected_feats)) > 1) {
       tmp <- data.table::data.table(selected_feats)
       tmp[, N := seq_len(.N), by = selected_feats]
@@ -550,8 +541,10 @@ plot_boxplot_num_panel <- function(dt_response,
     }
 
     feat_lbl_levels <- selected_feats
+    plot_parts <- vector("list", length(selected_feats))
 
-    for (selected_feat in selected_feats) {
+    for (fi in seq_along(selected_feats)) {
+      selected_feat <- selected_feats[[fi]]
 
       if (selected_feat %chin% available_feats) {
         # remove NA
@@ -564,7 +557,7 @@ plot_boxplot_num_panel <- function(dt_response,
         } else {
           # dummy data when all data is NA
           feat_lbl <- paste0(selected_feat, ": all NAs")
-          feat_lbl_levels[which(selected_feats == selected_feat)] <- feat_lbl
+          feat_lbl_levels[fi] <- feat_lbl
 
           tab_plot_tmp <- data.table::data.table(
             cellline_name = "",
@@ -579,7 +572,7 @@ plot_boxplot_num_panel <- function(dt_response,
       } else {
         # dummy data required for faceting
         feat_lbl <- paste0(selected_feat, ": all NAs")
-        feat_lbl_levels[which(selected_feats == selected_feat)] <- feat_lbl
+        feat_lbl_levels[fi] <- feat_lbl
 
         tab_plot_tmp <- data.table::data.table(
           cellline_name = "",
@@ -591,8 +584,9 @@ plot_boxplot_num_panel <- function(dt_response,
                              old = c("cellline_name", "selected_metric"),
                              new = c(cellline_name, selected_metric))
       }
-      tab_plot_all <- rbind(tab_plot_all, tab_plot_tmp)
+      plot_parts[[fi]] <- tab_plot_tmp
     }
+    tab_plot_all <- data.table::rbindlist(plot_parts, fill = TRUE)
 
     # prep value ranges for y-axis
     range_y <- range(tab_plot_all[!is.infinite(get(selected_metric)), ][[selected_metric]])
@@ -600,9 +594,8 @@ plot_boxplot_num_panel <- function(dt_response,
 
     # prep the number of items in each category
     tab_count_all <- tab_plot_all[!is.na(get(selected_metric)), .N, by = c("feat_val", "feat_lbl")]
-    tab_count_all_possible <- expand.grid(feat_val = unique(tab_plot_all$feat_val),
-                                          feat_lbl = unique(tab_plot_all$feat_lbl),
-                                          stringsAsFactors = FALSE)
+    tab_count_all_possible <- data.table::CJ(feat_val = unique(tab_plot_all$feat_val),
+                                              feat_lbl = unique(tab_plot_all$feat_lbl))
     # fill lacking labels
 
     if (NROW(tab_count_all) < NROW(tab_count_all_possible)) {
@@ -683,17 +676,13 @@ plot_boxplot_meta <- function(dt_response,
   checkmate::assert_number(max_x_lbl_length, lower = 5)
   checkmate::assert_flag(with_inf)
   boxplot_fill <-
-    gDRutils::get_settings_from_json("BOXPLOT_FILL",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("BOXPLOT_FILL")
   hline_color <-
-    gDRutils::get_settings_from_json("HLINE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("HLINE_COLOR")
   jitter_poinst_color <-
-    gDRutils::get_settings_from_json("JITTER_POINST_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("JITTER_POINST_COLOR")
   edge_color <-
-    gDRutils::get_settings_from_json("EDGE_COLOR",
-                                     system.file(package = "gDRplots", "settings.json"))
+    .get_setting("EDGE_COLOR")
 
   selected_metric <- setdiff(names(dt_response),
                              c(cellline_name, "rId", "cId"))
@@ -886,9 +875,8 @@ plot_volcano_assoc_panel <- function(dt_response,
            sprintf("%s__%s", selected_metric, selected_feat_meta_col),
            sprintf("%s__%s\n%s", selected_metric, selected_feat_meta_col, obj_assoc[["condition_info"]]))
 
-  panel <- ggpubr::annotate_figure(
-    ggpubr::ggarrange(plotlist = list(plt_vol, plt_side), widths = c(1, 1)),
-    top = panel_title)
+  panel <- patchwork::wrap_plots(plt_vol, plt_side, ncol = 2, widths = c(1, 1)) +
+    patchwork::plot_annotation(title = panel_title)
 
   # final
   return(list(assoc_data = assoc_data,