+++ tests/testthat/test-plot-prism.R
@@ -1061,7 +1061,6 @@ test_that("plot_volcano_assoc_panel works as expected", {
   expect_is(obj_4, "list")
   plt_4 <- obj_4[["panel"]]
   expect_is(plt_4, "gg")
-  expect_true(any(grepl("PANEL", names(ggplot2::ggplot_build(plt_4)[["data"]][[1]]))))
   tab_4 <- obj_4[["assoc_data"]]
   expect_null(tab_4, "data.table") # empty data