+++ NEWS.md
@@ -1,4 +1,6 @@
 ## gDRplots 0.0.117 - 2026-06-29
+* replace ggpubr with patchwork for combo heatmap panels and fitting QC
+* add gtable/grob support in save_plot
 * fix combo curve plot crash for missing normalization data
 
 ## gDRplots 0.0.116 - 2026-06-23
@@ -8,7 +10,7 @@
 * update dependencies
 
 ## gDRplots 0.0.114 - 2026-06-22
-* update pacakge vignette
+* update package vignette
 
 ## gDRplots 0.0.113 - 2026-06-17
 * add auto-orient combo dose-response axes by dose count